+++ src/util.rs
@@ -1,7 +1,7 @@
 use std::num::Int;
 use std::iter::repeat;
 
-pub const BITS_PER_WORD: uint = 32;
+pub const BITS_PER_WORD: usize = 32;
 
 // An M x N matrix of bits
 // The representation is in row-major form.
@@ -9,14 +9,14 @@ pub const BITS_PER_WORD: uint = 32;
 pub struct Bitmat
 {
 	pub data: Vec<u32>,			// contains all bits in matrix, in row-major form, with padding at end of row
-	pub rows: uint,				// number of rows
-	pub cols: uint,				// number of columns
-	pub rowsize: uint			// number of u32 elements per row
+	pub rows: usize,				// number of rows
+	pub cols: usize,				// number of columns
+	pub rowsize: usize			// number of u32 elements per row
 }
 
 impl Bitmat
 {
-	pub fn new(rows: uint, cols: uint) -> Bitmat
+	pub fn new(rows: usize, cols: usize) -> Bitmat
 	{
 		let rowsize = word_size(cols);
 		let total = rowsize * rows;
@@ -32,14 +32,14 @@ impl Bitmat
 
     // note: r is a word index, not a bit/row index, for the start of a row
 /*
-	pub fn bit(&self, r: uint, n: uint) -> bool {
+	pub fn bit(&self, r: usize, n: usize) -> bool {
     	(((self.data[r + (n >> 5)]) >> (n & 31)) & 1u32) != 0
 	}
 */
 
 /*
     // note: r is a word index, not a bit/row index, for the start of a row
-	pub fn set_bit(&mut self, r: uint, c: uint) {
+	pub fn set_bit(&mut self, r: usize, c: usize) {
         assert!(r < self.rows * self.rowsize);
         assert!(c < self.cols);
     	self.data[r + (c >> 5)] |= 1u32 << (c & 31);
@@ -48,13 +48,13 @@ impl Bitmat
 
     // sets an entry to 1, given a row/column index.
     // note that r and c are row and column indices, not word offsets.
-    pub fn set(&mut self, r: uint, c: uint) {
+    pub fn set(&mut self, r: usize, c: usize) {
         assert!(r < self.rows);
         assert!(c < self.cols);
         self.data[r * self.rowsize + (c >> 5)] |= 1u32 << (c & 31);
     }
 
-    pub fn get(&self, r: uint, c: uint) -> bool {
+    pub fn get(&self, r: usize, c: usize) -> bool {
         assert!(r < self.rows);
         assert!(c < self.cols);
         (self.data[r * self.rowsize + (c >> 5)] & (1u32 << (c & 31))) != 0u32
@@ -63,7 +63,7 @@ impl Bitmat
     // note: r is a row index, not a word index, for identifying a row.
     // This method returns an iterator which gives the positions of all
     // columns within a particular row 'r' where the row/col = 1.
-    pub fn iter_ones_in_row<'a>(&'a self, r: uint) -> BitMaskIterator<'a> {
+    pub fn iter_ones_in_row<'a>(&'a self, r: usize) -> BitMaskIterator<'a> {
         assert!(r < self.rows);
         // debug!("Bitmat.iter_ones_in_row: rows={} cols={} rowsize={} r={} (rowsize * r)={}", self.rows, self.cols, self.rowsize, r, self.rowsize * r);
         let rpos = self.rowsize * r;
@@ -84,18 +84,18 @@ impl Bitmat
 
 pub struct BitmatIterOnes<'a>
 {
-    row: uint,          // index of current row
-    col: uint,          // index of current col
+    row: usize,          // index of current row
+    col: usize,          // index of current col
     current: u32,       // contents of current word
     xormask: u32,       // xor mask used when loading current, used to select true/false
     mat: &'a Bitmat
 }
 
 impl<'a> Iterator for BitmatIterOnes<'a> {
-    type Item = (uint, uint);
+    type Item = (usize, usize);
     // This could be made faster with a "find first set bit" intrinsic.
-    fn next(&mut self) -> Option<(uint, uint)> {
-        const LOW_MASK: uint = (1 << BITS_PER_WORD) - 1;
+    fn next(&mut self) -> Option<(usize, usize)> {
+        const LOW_MASK: usize = (1 << BITS_PER_WORD) - 1;
 
         loop { // loop over rows
             while self.col < self.mat.cols { // loop over columns
@@ -131,7 +131,7 @@ impl<'a> Iterator for BitmatIterOnes<'a> {
     }
 }
 
-pub fn word_size(n: uint) -> uint
+pub fn word_size(n: usize) -> usize
 {
     (n + (BITS_PER_WORD - 1)) / BITS_PER_WORD
 }
@@ -141,16 +141,16 @@ pub struct BitMaskIterator<'a>
     words: &'a [u32],
     current: u32,       // contains the bits that we are currently reading
     xormask: u32,       // xor mask used when loading current, used to select true/false
-    nbits: uint,        // number of bits remaining in entire sequence
-    bitpos: uint,       // current bit position
+    nbits: usize,        // number of bits remaining in entire sequence
+    bitpos: usize,       // current bit position
 }
 
 impl<'a> Iterator for BitMaskIterator<'a>
 {
-    type Item = uint;
+    type Item = usize;
     // This could be made faster with a "find first set bit" intrinsic.
-    fn next(&mut self) -> Option<uint> {
-        const LOW_MASK: uint = (1 << BITS_PER_WORD) - 1;
+    fn next(&mut self) -> Option<usize> {
+        const LOW_MASK: usize = (1 << BITS_PER_WORD) - 1;
 
         while self.bitpos < self.nbits {
             // Are we at a word boundary?  If so, we need to load 'current'.
@@ -178,7 +178,7 @@ impl<'a> Iterator for BitMaskIterator<'a>
 }
 
 /// Iterates the indices of all of the bits set to 1 in a given bit vector.
-pub fn bit_vector_iter_ones<'a>(words: &'a [u32], nbits: uint) -> BitMaskIterator<'a> {
+pub fn bit_vector_iter_ones<'a>(words: &'a [u32], nbits: usize) -> BitMaskIterator<'a> {
     assert!(words.len() >= ((nbits + BITS_PER_WORD - 1) / BITS_PER_WORD));
     BitMaskIterator {
         words: words,
@@ -192,11 +192,11 @@ pub fn bit_vector_iter_ones<'a>(words: &'a [u32], nbits: uint) -> BitMaskIterato
 
 pub struct Bitv32 {
     pub data: Vec<u32>,
-    pub nbits: uint
+    pub nbits: usize
 }
 
 impl Bitv32 {
-    pub fn from_elem(n: uint, value: bool) -> Bitv32 {
+    pub fn from_elem(n: usize, value: bool) -> Bitv32 {
         let w = if value { !0u32 } else { 0u32 };
         let nwords = (n + BITS_PER_WORD - 1) / BITS_PER_WORD;
         Bitv32 {
